+++ apps/expo/app/server/[id]/books/[id]/index.tsx
@@ -12,13 +12,7 @@ import { useLocalSearchParams, useNavigation, useRouter } from 'expo-router'
 import { ChevronLeft } from 'lucide-react-native'
 import { useCallback, useLayoutEffect, useState } from 'react'
 import { Platform, Pressable, View } from 'react-native'
-import Animated, {
-	Extrapolation,
-	interpolate,
-	useAnimatedRef,
-	useAnimatedStyle,
-	useScrollOffset,
-} from 'react-native-reanimated'
+import Animated from 'react-native-reanimated'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import TImage from 'react-native-turbo-image'
 
@@ -29,8 +23,9 @@ import {
 	DescriptionSection,
 	DownloadButton,
 	IdentifiersSheet,
+	InfoRow,
+	useOverviewAnimations,
 } from '~/components/book/overview'
-import { InfoRow } from '~/components/book/overview'
 import { ThumbnailImage } from '~/components/image'
 import { MetadataBadgeSection } from '~/components/overview'
 import RefreshControl from '~/components/RefreshControl'
@@ -212,16 +207,7 @@ export default function Screen() {
 		}
 	}, [navigation, book, bookID])
 
-	const animatedScrollRef = useAnimatedRef<Animated.ScrollView>()
-	const scrollOffset = useScrollOffset(animatedScrollRef)
-
-	const parallaxStyle = useAnimatedStyle(() => {
-		return {
-			transform: [
-				{ translateY: interpolate(scrollOffset.value, [0, 200], [0, 100], Extrapolation.EXTEND) },
-			],
-		}
-	})
+	const { animatedScrollRef, parallaxStyle } = useOverviewAnimations()
 
 	if (!book) return null
