+++ apps/expo/components/book/overview/index.ts
@@ -4,3 +4,4 @@ export { default as DownloadButton } from './DownloadButton'
 export { default as IdentifiersSheet } from './IdentifiersSheet'
 export { default as InfoRow } from './InfoRow'
 export { default as LongValue } from './LongValue'
+export { useOverviewAnimations } from './useOverviewAnimations'
