+++ apps/expo/components/bookClub/PreviewBookSheet.tsx
@@ -3,19 +3,14 @@ import { useGraphQL, useSDK } from '@stump/client'
 import { graphql, PreviewBookSheetQuery } from '@stump/graphql'
 import { forwardRef, useImperativeHandle, useRef } from 'react'
 import { Platform, View } from 'react-native'
-import Animated, {
-	Extrapolation,
-	interpolate,
-	useAnimatedRef,
-	useAnimatedStyle,
-	useScrollOffset,
-} from 'react-native-reanimated'
+import Animated from 'react-native-reanimated'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import TImage from 'react-native-turbo-image'
 
 import { IS_IOS_24_PLUS, useColors } from '~/lib/constants'
 import { usePreferencesStore } from '~/stores'
 
+import { useOverviewAnimations } from '../book/overview'
 import { ThumbnailImage } from '../image'
 import { Button, Heading, Text } from '../ui'
 
@@ -117,16 +112,7 @@ function BookContent({ book, onConfirmAddBook }: BookContentProps) {
 
 	const thumbnailRatio = usePreferencesStore((store) => store.thumbnailRatio)
 
-	const animatedScrollRef = useAnimatedRef<Animated.ScrollView>()
-	const scrollOffset = useScrollOffset(animatedScrollRef)
-
-	const parallaxStyle = useAnimatedStyle(() => {
-		return {
-			transform: [
-				{ translateY: interpolate(scrollOffset.value, [0, 200], [0, 100], Extrapolation.EXTEND) },
-			],
-		}
-	})
+	const { animatedScrollRef, parallaxStyle } = useOverviewAnimations()
 
 	const { url: uri, metadata: placeholderData } = book.thumbnail
